refactor(core): active-messaging consolidation — handler-reuse, coherence model-fold, OoO extraction
Bring the runtime into conformance with the active-messaging design: the wire
RX dispatcher (MSG_*) and the OoO drain replay converge on a single
arts_handler_* per message via one unified engine, and the coherence subsystem
is folded per memory model and regrouped. Also fixes a latent LRC
ownership-transfer race and clears the tree's clang-tidy/clang-format backlog.

Handler-reuse + OoO engine
- Extract the out-of-order defer/replay engine to core/ooo.{c,h}: a single
  arts_ooo_dispatch_or_defer does the per-call lookup-acquire and either runs
  g_ooo_table[kind](item, args) or pushes the payload onto the slot's Treiber
  list. Handler bodies are pure (item, args) — no lookup/acquire/push.
- Converge the dispatcher: satisfy / add_dependence / destroy route straight
  through dispatch_or_defer (not the public API); MSG_SHUTDOWN -> a passive
  arts_handler_shutdown. Delete arts_coh_home_lookup_or_defer and the
  ooo_h_db_* packet-rebuild shims.
- ooo_kind: #ifdef ARTS_MEMORY_MODEL_* arms (per-model OOO_DB_*), drop the dead
  OOO_HANDLE_READY_EDT, fold OOO_EDT_SATISFY_SLOT_PTR into the mode-discriminated
  satisfy kind.

Coherence model-fold (Option B) + regroup
- Abolish the arts_coh_model_* 15-hook vtable seam (delete db_coherence_model.h);
  each per-model TU now defines arts_handler_db_* directly (CMake link-selects
  rc.c / lrc.c / lc.c; release.c is the RC+LRC family unit).
- Regroup db_coherence_*.{c,h} into core/coherence/ and arts/coherence/
  (coherence.c, rc.c, lrc.c, lc.c, release.c, buffer.c, home.c, senders.c,
  handlers.c; coherence.h / types.h / buffer.h / handlers.h / home.h).

Naming
- arts_remote_* -> arts_msg_* (wire packets) / arts_transport_* (plumbing) /
  arts_launcher_*; mark_delete -> set_destroyed.
- Purge the coh brand: arts_coh_* -> coh_ -> arts_db_* / static. The only
  arts_ prefixes left are API (arts_*), handlers (arts_handler_*), and the
  arts_db_acquire_all driver.

Coherence correctness (LRC ownership transfer)
- Make the LRC INVALIDATE writer_count commutative + signed like RC (additive
  sentinel, signed decrement firing only on the positive->0 edge) so a
  TRANSFER(+1)/INVALIDATE(-1) pair reordered to the same rank under
  receiver_threads>1 can't lose the transfer.
- Add a drain guard around the GRANT/TRANSFER waiter drain so the per-waiter
  count-up cannot transiently reach 0 mid-drain and ship the transfer before
  this rank's queued acquirers are counted (stranded-acquirer hang).
- New stress test tests/coherence_lrc_ownership_reorder.c (rt=2).

Lint
- clang-tidy (strict config, WarningsAsErrors) and clang-format (LLVM) now clean
  across libs/, tests/, and benchmarks/ (examples excluded): (void) on discarded
  returns, braces, math parens, redundant casts, const params, snake_case,
  parameter-name alignment. Behavior-preserving.

Verified across all three DB-coherence build models (RC/LRC/LC, GPU on):
clean build, single_node 85/85, multinode, coherence stress, OCR harness
Tier A/B/M; clang-tidy 0 / clang-format 0 over in-scope code.

+++ tests/multinode_event.c
@@ -100,24 +100,30 @@ void main_edt(uint32_t paramc, const uint64_t *paramv, uint32_t depc,
   // Test 1: Event on node 0, satisfied from node 1.  Default hint = ONCE
   // (latch=1, auto_destroy=true), home rank 0 by default.
   arts_guid_t ev1 = arts_event_create(NULL);
-  arts_guid_t done1 = arts_edt_create(event_done, 0, NULL, 1, &(arts_edt_hint_t){.rank = 0, .finish_event = fe});
+  arts_guid_t done1 =
+      arts_edt_create(event_done, 0, NULL, 1,
+                      &(arts_edt_hint_t){.rank = 0, .finish_event = fe});
   arts_add_dependence(ev1, done1, 0, DB_MODE_RW);
 
   uint64_t ev_param = (uint64_t)ev1;
-  arts_edt_create(remote_satisfier, 1, &ev_param, 0, &(arts_edt_hint_t){.rank = 1, .finish_event = fe});
+  arts_edt_create(remote_satisfier, 1, &ev_param, 0,
+                  &(arts_edt_hint_t){.rank = 1, .finish_event = fe});
 
   // Test 2: Event with latch = total_nodes, each node satisfies once.
   arts_event_hint_t fan_in_hint = ARTS_EVENT_HINT_DEFAULTS;
-  fan_in_hint.latch = total;
+  fan_in_hint.latch = (int32_t)total;
   fan_in_hint.rank = 0;
   arts_guid_t ev2 = arts_event_create(&fan_in_hint);
   uint64_t total_param = (uint64_t)total;
-  arts_guid_t done2 = arts_edt_create(fan_in_done, 1, &total_param, 1, &(arts_edt_hint_t){.rank = 0, .finish_event = fe});
+  arts_guid_t done2 =
+      arts_edt_create(fan_in_done, 1, &total_param, 1,
+                      &(arts_edt_hint_t){.rank = 0, .finish_event = fe});
   arts_add_dependence(ev2, done2, 0, DB_MODE_RW);
 
   for (unsigned int r = 0; r < total; r++) {
     uint64_t param = (uint64_t)ev2;
-    arts_edt_create(node_satisfier, 1, &param, 0, &(arts_edt_hint_t){.rank = r, .finish_event = fe});
+    arts_edt_create(node_satisfier, 1, &param, 0,
+                    &(arts_edt_hint_t){.rank = r, .finish_event = fe});
   }
 }
